fix(hwp): isolate parser from optional dependencies
Move balanced table scanning into a dependency-free module so repository
CI can typecheck its regression test without installing HWP-only packages.
